Handle illegal operation cases
Handle the following illegal operations cases:

- Trying to add/remove a product on a paid cart
- Trying to add a product to a cart that already has that product

All of these operations are not valid on the current architechture of the system

+++ src/main/kotlin/com/jumarket/selfcheckout/services/impl/CartService.kt
@@ -3,10 +3,14 @@ package com.jumarket.selfcheckout.services.impl
 import com.jumarket.selfcheckout.entities.Cart
 import com.jumarket.selfcheckout.entities.CartItem
 import com.jumarket.selfcheckout.enumerations.PaymentMethod
+import com.jumarket.selfcheckout.exceptions.CartAlreadyPaidException
 import com.jumarket.selfcheckout.exceptions.CartNotFoundException
+import com.jumarket.selfcheckout.exceptions.ProductAlreadyOnCartException
 import com.jumarket.selfcheckout.repositories.CartItemRepository
 import com.jumarket.selfcheckout.repositories.CartRepository
 import com.jumarket.selfcheckout.services.ICartService
+import org.springframework.data.domain.Example
+import org.springframework.data.domain.ExampleMatcher
 import org.springframework.stereotype.Service
 
 @Service
@@ -26,10 +30,13 @@ class CartService(
     }
 
     override fun addCartItem(cartItem: CartItem) {
+        if (cartItem.cart!!.paid) throw CartAlreadyPaidException("Can not add itens to a paid cart")
+        if (isProductOnCart(cartItem)) throw ProductAlreadyOnCartException("Product ${cartItem.product?.productName} already on cart")
         cartItemRepository.save(cartItem)
     }
 
     override fun removeCartItem(cartItem: CartItem) {
+        if (cartItem.cart!!.paid) throw CartAlreadyPaidException("Can not remove itens of a paid cart")
         cartItemRepository.delete(cartItem)
     }
 
@@ -39,5 +46,12 @@ class CartService(
             paymentMethod = method
         }
         cartRepository.save(cart)
+    }
+
+    fun isProductOnCart(cartItem: CartItem): Boolean {
+        val matcher = ExampleMatcher.matching()
+            .withIgnorePaths("quantity", "totalPrice")
+        val cartItemExample = Example.of(CartItem(cart = cartItem.cart, product = cartItem.product), matcher)
+        return cartItemRepository.findOne(cartItemExample).isPresent()
     }
 }
